feat: Allow searching for devices that are not initially populated
The RegistryManager#GetDeviceAsync documentation says that it returns
"an approximation, not a definitive list" of the devices present in the
IoT Hub.  That's why not all devices appear even when the max is bumped
up.

To allow the tool to still be useful in these scenarios, the filter text
box will now check to see if the filter matches a device id, and if so
populate that device even if it wsn't initially in the list.